chore: moved `jx step bdd` to a separate package
